Exercício 1.a,b,c
