+++ semana19/aula56-criptografia-user_roles/template-backend/src/app.ts
@@ -1,0 +1,21 @@
+import express, { Request, Response } from "express";
+import dotenv from "dotenv";
+import { AddressInfo } from "net";
+
+dotenv.config();
+
+const app = express();
+
+app.use(express.json());
+
+const server = app.listen(process.env.PORT || 3003, () => {
+  if (server) {
+    const address = server.address() as AddressInfo;
+    console.log(`Server is running in http://localhost:${address.port}`);
+  } else {
+    console.error(`Failure upon starting server.`);
+  }
+});
+
+
+export default app;
