+++ semana19/aula56-criptografia-user_roles/template-backend/src/index.ts
@@ -1,18 +1,22 @@
-import express, { Request, Response } from "express";
-import dotenv from "dotenv";
-import { AddressInfo } from "net";
+import { generetorHash, compare } from './services/generetorHash';
+import app from "./app";
 
-dotenv.config();
 
-const app = express();
 
-app.use(express.json());
+async function cripto(){
+  const hm = await generetorHash("Vamos encriptografar");
 
-const server = app.listen(process.env.PORT || 3003, () => {
-  if (server) {
-    const address = server.address() as AddressInfo;
-    console.log(`Server is running in http://localhost:${address.port}`);
-  } else {
-    console.error(`Failure upon starting server.`);
-  }
-});
+  console.log(hm)
+}
+
+// cripto();
+
+
+
+async function comparar(){
+  const c = await compare("Vamos encriptografar", "$2a$12$XIMVnldegvA00BxJbc9ImOQ15jxijHZgGICtDFVRHiyxoJoLmHcsW");
+
+  console.log(c)
+}
+
+comparar();
